Fixes #385 - Occasional bugs in bulk scanning parts
