+++ Binner/Binner.Web/ClientApp/index.html
@@ -12,7 +12,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Roboto+Mono:wght@300&family=Source+Code+Pro&display=swap" rel="stylesheet">
     <title>Binner.Web</title>
 </head>
-  <body>
+  <body tabindex="1">
     <noscript>
       You need to enable JavaScript to run this app.
     </noscript>
