+++ Binner/Binner.Web/ClientApp/src/components/modals/BulkScanModal.css
