+++ Binner/Binner.Web/ClientApp/src/pages/Inventory.css
@@ -292,3 +292,50 @@
 .inventory .ui.secondary.segment.part-location {
   padding-top: 5px;
 }
+
+body:focus #bulkScanModal.ui.modal {
+  border: 20px solid #0c0;
+}
+
+#bulkScanModal.ui.modal {
+  transition-property: border;
+  transition-duration: 500ms;
+  transition-timing-function: ease;
+}
+#bulkScanModal.ui.modal .focusCheck {
+  transition-property: border;
+  transition-duration: 500ms;
+  transition-timing-function: ease;
+  border: 4px dashed transparent;
+}
+
+#bulkScanModal.ui.modal:focus .focusCheck,
+#bulkScanModal.ui.modal:focus-within .focusCheck,
+#bulkScanModal.ui.modal:focus-visible .focusCheck {
+  visibility: hidden !important;
+}
+
+#bulkScanModal.ui.modal:not(:focus,:focus-within,:focus-visible) .focusCheck, 
+#bulkScanModal.ui.modal.visible:not(:focus,:focus-within,:focus-visible) .focusCheck {
+  visibility: visible !important;
+  border: 4px dashed #f00;
+  text-align: center;
+  font-size: 1.5em;
+  color: #fff;
+}
+
+#bulkScanModal.ui.modal {
+  border: 20px solid #f00;
+}
+
+#bulkScanModal.ui.modal:focus {
+  border: 20px solid #0c0 !important;
+}
+
+#bulkScanModal.ui.modal:focus-within {
+  border: 20px solid #060;
+}
+
+#bulkScanModal.ui.modal:focus-visible {
+  border: 20px solid #060;
+}
