+++ Binner/Library/Binner.Global.Common/UnauthorizedException.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace Binner.Global.Common
+{
+    public class UnauthorizedException : Exception
+    {
+        public UnauthorizedException() : base("User is not logged in")
+        {
+
+        }
+        public UnauthorizedException(string message) : base(message)
+        {
+        }
+    }
+}
